+++ pyobjc-framework-CallKit/PyObjCTest/test_cxcalldirectory.py
@@ -1,0 +1,8 @@
+from PyObjCTest.TestSupport import TestCase
+import CallKit
+import sys
+
+
+class TestCXCallDirectory(TestCase):
+    def test_constants(self):
+        self.assertEqual(CallKit.CXCallDirectoryPhoneNumberMax, sys.maxsize - 1)