+++ pyobjc-framework-CallKit/PyObjCTest/test_cxcallobserver.py
@@ -1,0 +1,8 @@
+from PyObjCTest.TestSupport import TestCase
+import CallKit  # noqa: F401
+import objc
+
+
+class TestCXCallObserver(TestCase):
+    def test_protocols(self):
+        objc.protocolNamed("CXCallObserverDelegate")