+++ pyobjc-framework-CallKit/PyObjCTest/test_cxsetheldcallaction.py
@@ -1,0 +1,8 @@
+from PyObjCTest.TestSupport import TestCase
+import CallKit
+
+
+class TestCXSetHeldCallAction(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(CallKit.CXSetHeldCallAction.isOnHold)
+        self.assertArgIsBOOL(CallKit.CXSetHeldCallAction.setOnHold_, 0)