+++ pyobjc-framework-CallKit/PyObjCTest/test_cxsetmutedcallaction.py
@@ -1,0 +1,8 @@
+from PyObjCTest.TestSupport import TestCase
+import CallKit
+
+
+class TestCXSetMutedCallAction(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(CallKit.CXSetMutedCallAction.isMuted)
+        self.assertArgIsBOOL(CallKit.CXSetMutedCallAction.setMuted_, 0)