+++ pyobjc-framework-CallKit/PyObjCTest/test_cxtransaction.py
@@ -1,0 +1,7 @@
+from PyObjCTest.TestSupport import TestCase
+import CallKit
+
+
+class TestCXTransaction(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(CallKit.CXTransaction.isComplete)